auth-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
passlib[bcrypt]==1.7.4 # Para hashing de contraseñas
python-jose[cryptography]==3.3.0 # Para JWT
SQLAlchemy==2.0.30 # ORM para la base de datos